action menu tutorial tsx
