--- react/example-customer-account--loyalty-app--react/extensions/customer-account-loyalty-extension/locales/fr.json
@@ -1,3 +1,0 @@
-{
-  "welcome": "Bienvenue à la cible {{target}} de l'extension d'interface utilisateur du compte client!"
-}
