+++ react/example-customer-account--order-action-menu--react/extensions/action-menu-extension-react/.gitignore
@@ -1,0 +1,1 @@
+/dist
