+++ react/example-customer-account--order-action-menu--react/extensions/action-menu-extension-react/README.md
@@ -1,0 +1,3 @@
+# Customer account UI Extension
+
+Allow developers to build extensions on customer account.
