+++ react/example-customer-account--order-action-menu--react/extensions/action-menu-extension-react/locales/en.default.json
@@ -1,0 +1,8 @@
+{
+  "report": "Report a problem",
+  "menuAction": {
+    "title": "Report a problem",
+    "cancel": "Cancel",
+    "submit": "Report"
+  }
+}
