+++ react/example-customer-account--order-action-menu--react/extensions/action-menu-extension-react/shopify.extension.toml
@@ -1,0 +1,14 @@
+api_version = "2023-07"
+
+[[extensions]]
+type = "ui_extension"
+name = "action-menu-extension-react"
+handle = "action-menu-extension-react"
+
+[[extensions.targeting]]
+module = "./src/MenuActionExtension.tsx"
+target = "customer-account.order.action.menu-item.render"
+
+[[extensions.targeting]]
+module = "./src/MenuActionItemExtension.tsx"
+target = "customer-account.order.action.render"
