+++ react/example-customer-account--order-action-menu--react/extensions/action-menu-extension-react/src/MenuActionItemExtension.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import {
   Button,
   CustomerAccountAction,
