+++ react/example-customer-account--order-action-menu--react/shopify.app.toml
@@ -1,0 +1,22 @@
+# Learn more about configuring your app at https://shopify.dev/docs/apps/tools/cli/configuration
+
+name = "action-menu-extension-react"
+client_id = "fc7a9a9a255d21a2effa007dafedf5e8"
+application_url = "https://shopify.dev/apps/default-app-home"
+embedded = true
+
+[access_scopes]
+# Learn more at https://shopify.dev/docs/apps/tools/cli/configuration#access_scopes
+scopes = ""
+
+[auth]
+redirect_urls = [ "https://shopify.dev/apps/default-app-home/api/auth" ]
+
+[webhooks]
+api_version = "2023-10"
+
+[pos]
+embedded = false
+
+[build]
+dev_store_url = "development-store-1.shopify.tutorials-setup.oluwatimi-owoturo.us.spin.dev"
